requirements.txt
openai
python-dotenv
git+https://github.com/openai/swarm.git
requests
beautifulsoup4